Add tsconfig and rename the files
