+++ pages/App.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import CarForm from "../components/CarForm";
+import CarList from "../components/CarList";
+import CarSearch from "../components/CarSearch";
+import CarValue from "../components/CarValue";
+
+const App = () => {
+  return (
+    <>
+      <CarForm />
+      <CarList />
+      <CarSearch />
+      <CarValue />
+    </>
+  );
+};
+
+export default App;
